Remove the upper limit on Ruby version
If a gem has an upper limit on the Ruby version it depends on, we need to update that limit every time a new Ruby version is released, which is a hassle. Since Ruby prioritizes compatibility between versions, things should generally work as expected even after an upgrade. The disadvantages of setting an upper version limit outweigh the benefits. So, I’m removing the upper limit on the Ruby version dependency.